+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Interfaces/IContentPartGraphSyncer.cs
@@ -4,7 +4,7 @@ using DFC.ServiceTaxonomy.Neo4j.Commands.Interfaces;
 using Neo4j.Driver;
 using OrchardCore.ContentManagement.Metadata.Models;
 
-namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces
 {
     public interface IContentPartGraphSyncer
     {
