+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Interfaces/IGraphSyncPartIdProperty.cs
@@ -1,4 +1,4 @@
-namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces
 {
     public interface IGraphSyncPartIdProperty
     {
