+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Interfaces/IGraphSyncer.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using DFC.ServiceTaxonomy.Neo4j.Commands.Interfaces;
 using Newtonsoft.Json.Linq;
 
-namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces
 {
     public interface IGraphSyncer
     {
