+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/TitlePartGraphSyncer.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
 using DFC.ServiceTaxonomy.Neo4j.Commands.Interfaces;
 using Neo4j.Driver;
 using OrchardCore.ContentManagement.Metadata.Models;
